+++ flow/src/main/scala/org/alephium/flow/core/BlockPool.scala
@@ -16,6 +16,8 @@
 
 package org.alephium.flow.core
 
+import java.math.BigInteger
+
 import org.alephium.io.IOResult
 import org.alephium.protocol.BlockHash
 import org.alephium.protocol.model.Block
@@ -29,7 +31,7 @@ trait BlockPool extends BlockHashPool {
   def getBlock(hash: BlockHash): IOResult[Block]
 
   // Assuming the block is verified
-  def add(block: Block, weight: BigInt): IOResult[Unit]
+  def add(block: Block, weight: BigInteger): IOResult[Unit]
 
   def getBlocksAfter(locator: BlockHash): IOResult[AVector[Block]] = {
     for {
@@ -40,7 +42,7 @@ trait BlockPool extends BlockHashPool {
 
   def getHeight(block: Block): IOResult[Int] = getHeight(block.hash)
 
-  def getWeight(block: Block): IOResult[BigInt] = getWeight(block.hash)
+  def getWeight(block: Block): IOResult[BigInteger] = getWeight(block.hash)
 
   // TODO: use ChainSlice instead of AVector[Block]
   def getBlockSlice(hash: BlockHash): IOResult[AVector[Block]] = {
